feat(cli): Rewrite CLI with async implementation
Instead of spawning one thread for each device, in addition to a thread
for watching them, this will use a single background thread to watch for
file descriptor readiness, and another to concurrently write to all
devices from the same thread.

The background thread which performs the flashing will also be spawned
during application init so that it is immediately ready to receive the
flash task as soon as it is available from the main thread.

Error handling has also been improved in this implementation.