chore: Version 1.1.0
The version numbers in the library, CLI, and GUI were all significantly
different. I've made them the same, which seems easier to keep track of,
unless there's a good reason to do otherwise.

This version can then be tagged, and we should then be able to submit
Popsicle to Flathub.